tests: Move TestAudioBackend to backends/audio.rs
