+++ tests/framework/src/backends.rs
@@ -1,3 +1,5 @@
+mod audio;
 mod navigator;
 
+pub use audio::TestAudioBackend;
 pub use navigator::TestNavigatorBackend;
