+++ tests/framework/src/runner.rs
@@ -5,15 +5,10 @@ use crate::options::ImageComparison;
 use crate::test::Test;
 use crate::test_ui::TestUiBackend;
 use anyhow::{anyhow, Result};
-use ruffle_core::backend::audio::{
-    swf, AudioBackend, AudioMixer, DecodeError, RegisterError, SoundHandle, SoundInstanceHandle,
-    SoundStreamInfo, SoundTransform,
-};
 use ruffle_core::backend::log::LogBackend;
 use ruffle_core::backend::navigator::NullExecutor;
 use ruffle_core::events::MouseButton as RuffleMouseButton;
 use ruffle_core::events::{KeyCode, TextControlCode as RuffleTextControlCode};
-use ruffle_core::impl_audio_mixer_backend;
 use ruffle_core::limits::ExecutionLimit;
 use ruffle_core::tag_utils::SwfMovie;
 use ruffle_core::{Player, PlayerBuilder, PlayerEvent};
@@ -28,41 +23,6 @@ use std::path::Path;
 use std::rc::Rc;
 use std::sync::{Arc, Mutex};
 use std::time::Duration;
-
-pub struct TestAudioBackend {
-    mixer: AudioMixer,
-    buffer: Vec<f32>,
-}
-
-impl Default for TestAudioBackend {
-    fn default() -> Self {
-        Self {
-            mixer: AudioMixer::new(Self::NUM_CHANNELS, Self::SAMPLE_RATE),
-            buffer: vec![],
-        }
-    }
-}
-
-impl TestAudioBackend {
-    const NUM_CHANNELS: u8 = 2;
-    const SAMPLE_RATE: u32 = 44100;
-}
-
-impl AudioBackend for TestAudioBackend {
-    impl_audio_mixer_backend!(mixer);
-    fn play(&mut self) {}
-    fn pause(&mut self) {}
-
-    fn set_frame_rate(&mut self, frame_rate: f64) {
-        let new_buffer_size =
-            ((Self::NUM_CHANNELS as u32 * Self::SAMPLE_RATE) as f64 / frame_rate).round() as usize;
-        self.buffer.resize(new_buffer_size, 0.0);
-    }
-    fn tick(&mut self) {
-        debug_assert!(!self.buffer.is_empty());
-        self.mixer.mix::<f32>(self.buffer.as_mut());
-    }
-}
 
 #[derive(Clone)]
 pub struct TestLogBackend {
